requirements.txt
python-debian>=0.1.37